Wrap DefaultConfigBuilder in static factory method
DefaultConfigBuilder is different from ConfigBuilder in that it takes
an app name and does not build the chain of logs. Simplify the API
by wrapping DefaultConfigBuilder into DefaultConfig's static factory
methods.
