+++ build.gradle
@@ -24,7 +24,7 @@ plugins {
 
 group = 'org.computelab'
 ext.name = 'config'
-version = '0.1.0'
+version = '0.2.0'
 version = staging ? version : version + '-SNAPSHOT'
 
 /*** Java ***/
