+++ src/test/java/org/computelab/config/DefaultConfigTest.java
@@ -10,42 +10,49 @@ import org.apache.commons.lang3.reflect.FieldUtils;
 import org.junit.Test;
 
 @SuppressWarnings("unchecked")
-public class DefaultConfigBuilderTest {
+public class DefaultConfigTest {
 
     @Test
-    public void testConstructor() throws IllegalAccessException {
+    public void testCreate() {
         final String appName = getClass().getSimpleName();
-        final DefaultConfigBuilder builder = new DefaultConfigBuilder(appName);
+        final Config config = DefaultConfig.create(appName);
+        assertNotNull(config);
+    }
+
+    @Test(expected=NullPointerException.class)
+    public void createAppNameCannotBeNull() {
+        DefaultConfig.create(null, "configFile");
+    }
+
+    @Test(expected=IllegalArgumentException.class)
+    public void createAppNameCannotBeEmpty() {
+        DefaultConfig.create("", "configFile");
+    }
+
+    @Test(expected=NullPointerException.class)
+    public void createConfigFileCannotBeNull() {
+        DefaultConfig.create("app", null);
+    }
+
+    @Test(expected=IllegalArgumentException.class)
+    public void createConfigFileCannotBeEmpty() {
+        DefaultConfig.create("app", "");
+    }
+
+    @Test
+    public void testBuilderConstructor() throws IllegalAccessException {
+        final String appName = getClass().getSimpleName();
+        final DefaultConfig.Builder builder = new DefaultConfig.Builder(appName, "app.properties");
         assertEquals(appName, FieldUtils.readField(
                 builder, "appName", true).toString());
         assertEquals("app.properties", FieldUtils.readField(
                 builder, "configFile", true).toString());
     }
 
-    @Test(expected=NullPointerException.class)
-    public void constructorAppNameCannotBeNull() {
-        new DefaultConfigBuilder(null, "configFile");
-    }
-
-    @Test(expected=IllegalArgumentException.class)
-    public void constructorAppNameCannotBeEmpty() {
-        new DefaultConfigBuilder("", "configFile");
-    }
-
-    @Test(expected=NullPointerException.class)
-    public void constructorConfigFileCannotBeNull() {
-        new DefaultConfigBuilder("app", null);
-    }
-
-    @Test(expected=IllegalArgumentException.class)
-    public void constructorConfigFileCannotBeEmpty() {
-        new DefaultConfigBuilder("app", "");
-    }
-
     @Test
     public void testBuild() throws IllegalAccessException {
         final String appName = getClass().getSimpleName();
-        final DefaultConfigBuilder builder = new DefaultConfigBuilder(appName);
+        final DefaultConfig.Builder builder = new DefaultConfig.Builder(appName, "app.properties");
         final Config config = builder.build();
         assertTrue(config instanceof ConfigChain);
         final List<Config> configs = (List<Config>)FieldUtils.readField(config, "configs", true);
@@ -58,7 +65,7 @@ public class DefaultConfigBuilderTest {
     @Test
     public void testBuildWithResource() throws IllegalAccessException {
         final String appName = getClass().getSimpleName();
-        final DefaultConfigBuilder builder = new DefaultConfigBuilder(appName, "test.properties");
+        final DefaultConfig.Builder builder = new DefaultConfig.Builder(appName, "test.properties");
         assertEquals("test.properties", FieldUtils.readField(
                 builder, "configFile", true).toString());
         final Config config = builder.build();
